modified architecture
